+++ src/QuickMarcEditor/QuickMarcEditorRows/QuickMarcEditorRows.js
@@ -77,7 +77,7 @@ const QuickMarcEditorRows = ({
   const indexOfNewRow = useRef(null);
   const newRowRef = useRef(null);
 
-  const { linkAuthority, unlinkAuthority, linkableBibFields } = useAuthorityLinking();
+  const { linkAuthority, unlinkAuthority, linkableBibFields, sourceFiles } = useAuthorityLinking();
 
   const isNewRow = useCallback((row) => {
     return !initialValues.records.find(record => record.id === row.id);
@@ -445,6 +445,7 @@ const QuickMarcEditorRows = ({
                       isLinked={recordRow._isLinked}
                       tag={recordRow.tag}
                       fieldId={recordRow.id}
+                      sourceFiles={sourceFiles}
                     />
                   )}
                   {canViewAuthorityRecord && (