+++ src/hooks/useAuthorityLinking/useAuthorityLinking.js
@@ -153,6 +153,7 @@ const useAuthorityLinking = () => {
     linkAuthority,
     unlinkAuthority,
     linkableBibFields,
+    sourceFiles,
   };
 };